Use bare `require` in conditional URL import
This commit changes the code that sets up URL parsing in Glimmer to use the bare, global `require` in Node.js environments, rather than importing from our `node-module` helper package.

The `node-module` package is a shim we emit in the AMD build that exposes the Node `module` and `require` globals which would otherwise be shadowed by AMD. However, this creates problems when compiling Ember to package formats where we want to ultimately rely on standard Node resolution semantics, since `node-module` is not a package that appears in `node_modules`.

Instead, we switch the code to use a bare `require` and rely on the existing `inject-node-globals` Babel transform to detect this usage of bare `require` and inject the appropriate AMD require, only in the AMD build.